+++ app/src/main/java/com/techcourse/support/web/adaptor/ManualHandlerAdaptor.java
@@ -1,0 +1,26 @@
+package com.techcourse.support.web.adaptor;
+
+import com.techcourse.support.web.mapping.ManualHandlerMapping;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import webmvc.org.springframework.web.servlet.ModelAndView;
+import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
+import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerAdaptor;
+import webmvc.org.springframework.web.servlet.view.JspView;
+
+public class ManualHandlerAdaptor implements HandlerAdaptor {
+
+    @Override
+    public boolean supports(Object handler) {
+        return handler instanceof ManualHandlerMapping;
+    }
+
+    @Override
+    public ModelAndView handle(Object handler, HttpServletRequest request, HttpServletResponse response)
+            throws Exception {
+        Controller controller = (Controller) handler;
+        String viewName = controller.execute(request, response);
+
+        return new ModelAndView(new JspView(viewName));
+    }
+}
